test: separate shared card preview coverage
