+++ playwright/supabase.spec.js
@@ -123,25 +123,6 @@ test("Supabase cards keep square previews while compact emails remain fully visi
   });
 });
 
-test("the shared card preview behavior scrolls a long template without growing its card", async ({
-  page,
-}) => {
-  await page.goto("index.html");
-
-  const image = page.getByAltText("Product Promotion Thumbnail");
-  await expect(image).toHaveCount(1);
-  const preview = image.locator("..");
-  const before = await preview.boundingBox();
-
-  await preview.hover();
-  await expect
-    .poll(() => image.evaluate((node) => node.style.transform))
-    .toContain("translateY(-");
-
-  const during = await preview.boundingBox();
-  expect(during?.height).toBeCloseTo(before?.height ?? 0, 2);
-});
-
 test("the reset password template opens with Supabase copy support", async ({
   page,
 }) => {
